+++ database/migrations/20180416090133_createRidesTable.js
@@ -10,7 +10,7 @@ exports.up = (knex, Promise) => {
     table.varchar('name', 255).notNullable().defaultTo('')
     table.varchar('strava_ride_id', 255).notNullable().defaultTo('')
     table.dateTime('rode_at').notNullable().defaultTo(knex.raw('now()'))
-    table.integer('distance').notNullable().defaultTo(0)
+    table.float('distance').notNullable().defaultTo(0)
     table.varchar('distance_unit', 20).notNullable()
     table.uuid('bike_id').notNullable()
     table.foreign('bike_id').references('id').inTable(bikesTable).onDelete('cascade')
